+++ project-addons/transportation/data/parameters.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data noupdate="1">
+        <record id="visiotech_web_tracking" model="ir.config_parameter">
+           <field name="key">url.visiotech.web.tracking</field>
+           <field name="value"></field>
+        </record>
+        <record id="visiotech_web_tracking_p" model="ir.config_parameter">
+           <field name="key">url.visiotech.web.tracking.pass</field>
+           <field name="value"></field>
+        </record>
+    </data>
+</odoo>
+
+
