+++ project-addons/transportation/models/__init__.py
@@ -1,0 +1,5 @@
+from . import res_partner
+from . import rotation
+from . import sale
+from . import stock
+from . import transportation
