+++ project-addons/transportation/views/res_partner_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <record id="res_partner_add_transport" model="ir.ui.view">
@@ -8,8 +8,10 @@
             <field name="inherit_id" ref="base.view_partner_form"/>
             <field name="arch" type="xml">
                 <field name="company_id" position="after">
-                    <field name="transporter_id" attrs="{'required': [('is_company', '=', True),('customer','=', True),('dropship', '=', False),('prospective', '=', False)]}"/>
-                    <field name="service_id" attrs="{'required': [('is_company', '=', True),('customer','=', True),('dropship', '=', False),('prospective', '=', False)]}"/>
+                    <!--field name="transporter_id" attrs="{'required': [('is_company', '=', True),('customer','=', True),('dropship', '=', False),('prospective', '=', False)]}"/>
+                    <field name="service_id" attrs="{'required': [('is_company', '=', True),('customer','=', True),('dropship', '=', False),('prospective', '=', False)]}"/-->
+                    <field name="transporter_id" attrs="{'required': [('is_company', '=', True),('customer','=', True)]}"/>
+                    <field name="service_id" attrs="{'required': [('is_company', '=', True),('customer','=', True)]}"/>
                 </field>
             </field>
         </record>
@@ -50,4 +52,4 @@
         </record>
 
     </data>
-</openerp>
+</odoo>
