+++ project-addons/transportation/views/sale_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
         <record id="sale_order_add_transporter" model="ir.ui.view">
             <field name="name">sale.order.add.transporter</field>
@@ -13,4 +13,4 @@
             </field>
         </record>
     </data>
-</openerp>
+</odoo>
