+++ project-addons/transportation/views/stock_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <record id="view_picking_withtransport_out_form" model="ir.ui.view">
@@ -7,7 +7,7 @@
             <field name="model">stock.picking</field>
             <field name="inherit_id" ref="stock.view_picking_form"/>
             <field name="arch" type="xml">
-                <xpath expr="//page[@string='Additional Info']/group[last()]" position="after">
+                <xpath expr="//page[@name='extra']/group[last()]" position="after">
                 <separator string="Carrier Information"/>
                 <group>
                     <group>
@@ -63,4 +63,4 @@
         </record>
 
     </data>
-</openerp>
+</odoo>
